osx: compile on OSX
This patch allows the full tree to build on OSX, but currently there are
a lot of segfaults, inconsistent uses of DARWIN/__APPLE__ defines, and
some of the semantic changes are likely wrong. We'll need to split this
up into a much longer patch series, and can probably start with the
minimal change set needed to make fuse work.

Use homebrew installed in its default location (/usr/local/) to install
dependencies, and configure with --without-libatomic-ops --with-libaio.

Use --without-tcmalloc: apparently broken with LLVM